# BACKEND/docs/swagger.yaml
definitions:
...
  controller.postInput:
    properties:
      categories_id:
        type: integer
      companies:
        type: string
      post_text:
        type: string
      post_title:
        type: string
      user_id:
        type: integer
    type: object
  controller.voteInput:
    properties:
      post_id:
        type: integer
      userID:
        type: integer
      vote_type:
        type: boolean
    type: object
...
  models.Post:
    properties:
      categories_id:
        type: integer
      companies:
        type: string
      created_at:
        type: string
      id:
        type: integer
      post_text:
        type: string
      post_title:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  models.Votes:
    properties:
      id:
        type: integer
      post_id:
        type: integer
      user_id:
        type: integer
      vote_type:
        type: boolean
    type: object
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
paths:
...
  /post:
    get:
      description: Get a list of Post.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Post'
            type: array
      summary: Get all Post.
      tags:
      - Post
    post:
      description: Creating a new Post.
      parameters:
      - description: the body to create a new Post
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controller.postInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Post'
      summary: Create New Post.
      tags:
      - Post
  /post/{id}:
    delete:
      description: Delete a Post by id.
      parameters:
      - description: post id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      summary: Delete one Post.
      tags:
      - Post
    get:
      description: Get a Post by text.
      parameters:
      - description: post text
        in: path
        name: text
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Post'
      summary: Get Post.
      tags:
      - Post
    put:
      description: Update Post by id.
      parameters:
      - description: post id
        in: path
        name: id
        required: true
        type: string
      - description: the body to update an Post
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controller.postInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Post'
      summary: Update Post.
      tags:
      - Post
...
  /vote:
    get:
      description: Get a list of Vote.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Votes'
            type: array
      summary: Get all Vote.
      tags:
      - Vote
    post:
      description: Creating a new Vote.
      parameters:
      - description: the body to create a new vote
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controller.voteInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Votes'
      summary: Create New Vote.
      tags:
      - Vote
  /vote/{id}:
    delete:
      description: Delete a Vote by id.
      parameters:
      - description: Vote id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      summary: Delete one Vote.
      tags:
      - Vote
swagger: "2.0"